+++ frame_forge/__init__.py
@@ -38,6 +38,7 @@ class GenerateImages:
         sub_size: int,
         sub_alignment: int,
         sub_color: str,
+        sub_secondary_color: str,
         sub_font_name: str,
         source_sub_title: str,
         release_sub_title: str,
@@ -68,6 +69,7 @@ class GenerateImages:
         self.sub_size = sub_size
         self.sub_alignment = sub_alignment
         self.sub_color = sub_color
+        self.sub_secondary_color = sub_secondary_color
         self.sub_font_name = sub_font_name
         self.source_sub_title = source_sub_title
         self.release_sub_title = release_sub_title
@@ -99,8 +101,12 @@ class GenerateImages:
         if self.sub_color:
             color = hex_to_bgr(self.sub_color)
 
+        secondary_color = "&H00000000"
+        if self.sub_secondary_color:
+            secondary_color = hex_to_bgr(self.sub_secondary_color)
+
         selected_sub_style = (
-            f"{self.sub_font_name},{self.sub_size},{color},&H00000000,&H00000000,&H00000000,"
+            f"{self.sub_font_name},{self.sub_size},{color},{secondary_color},&H00000000,&H00000000,"
             f"1,0,0,0,100,100,0,0,1,1,0,{self.sub_alignment},10,10,10,1"
         )
         sync_sub_base = (